refactor: Id 타입 Int64로 변경 - #138

+++ Projects/Domain/Goal/Interface/Sources/DTO/GoalCompleteResponseDTO.swift
@@ -9,7 +9,7 @@ import Foundation
 
 /// 목표 완료 응답 DTO입니다.
 public struct GoalCompleteResponseDTO: Decodable {
-    public let goalId: Int
+    public let goalId: Int64
     public let name: String
     public let goalStatus: String
     public let completedAt: String?
